CAMEL-16852 - Moved ServiceKeys model to camel-huaweicloud-common
CAMEL-16852 - Moved ServiceKeys model to camel-huaweicloud-common
