+++ components/camel-huawei/camel-huaweicloud-common/pom.xml
@@ -17,27 +17,29 @@
     limitations under the License.
 
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
         <groupId>org.apache.camel</groupId>
-        <artifactId>components</artifactId>
+        <artifactId>camel-huawei-parent</artifactId>
         <version>3.12.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>camel-huawei-parent</artifactId>
-    <packaging>pom</packaging>
-    <name>Camel :: Huawei Cloud :: Parent</name>
-    <description>Camel Huawei Cloud parent</description>
+    <artifactId>camel-huaweicloud-common</artifactId>
+    <packaging>jar</packaging>
+    <name>Camel :: Huawei Cloud :: Common</name>
+    <description>HuaweiCloud package containing shared classes</description>
 
-    <modules>
-		<module>camel-huaweicloud-common</module>
-        <module>camel-huaweicloud-functiongraph</module>
-        <module>camel-huaweicloud-iam</module>
-		<module>camel-huaweicloud-obs</module>
-        <module>camel-huaweicloud-smn</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-slf4j-impl</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 
 </project>
