+++ components/camel-huawei/camel-huaweicloud-functiongraph/src/test/java/org/apache/camel/InvokeFunctionServiceKeyFunctionalTest.java
@@ -17,9 +17,9 @@
 package org.apache.camel;
 
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.huaweicloud.common.models.ServiceKeys;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.constants.FunctionGraphProperties;
-import org.apache.camel.models.ServiceKeys;
 import org.apache.camel.test.junit4.CamelTestSupport;
 import org.junit.Ignore;
 import org.junit.Test;
