+++ components/camel-huawei/camel-huaweicloud-iam/pom.xml
@@ -73,5 +73,10 @@
       <artifactId>camel-test</artifactId>
       <scope>test</scope>
     </dependency>
+      <dependency>
+          <groupId>org.apache.camel</groupId>
+          <artifactId>camel-huaweicloud-common</artifactId>
+          <scope>compile</scope>
+      </dependency>
   </dependencies>
 </project>
